requirements.txt
appdirs==1.4.4
asgiref==3.3.1
beautifulsoup4==4.9.3
click==7.1.2
decorator==4.4.2
Django==3.1.7
django-environ==0.4.5
djangorestframework==3.12.2
Markdown==3.3.4
mccabe==0.6.1
mypy-extensions==0.4.3
pathspec==0.8.1
Pillow==8.1.2
psycopg2==2.8.6
psycopg2-binary==2.8.6
pytz==2021.1
regex==2021.3.17
six==1.15.0
sorl-thumbnail==12.7.0
soupsieve==2.2.1
sqlparse==0.4.1
toml==0.10.2
tornado==6.1
typed-ast==1.4.2
typing-extensions==3.7.4.3
